[weather][refactor] move weather related classes into its own subdir weather
