+++ xbmc/LangInfo.cpp
@@ -42,9 +42,9 @@
 #include "utils/log.h"
 #include "utils/StringUtils.h"
 #include "utils/URIUtils.h"
-#include "utils/Weather.h"
 #include "utils/XBMCTinyXML.h"
 #include "utils/XMLUtils.h"
+#include "weather/Weather.h"
 
 using namespace PVR;
 using namespace KODI::MESSAGING;
