+++ xbmc/weather/CMakeLists.txt
@@ -1,0 +1,7 @@
+set(SOURCES GUIWindowWeather.cpp
+            Weather.cpp)
+
+set(HEADERS GUIWindowWeather.h
+            Weather.h)
+
+core_add_library(weather)
