+++ xbmc/weather/GUIWindowWeather.h
