+++ xbmc/weather/Weather.cpp
@@ -24,7 +24,6 @@
 
 #include "addons/AddonManager.h"
 #include "addons/settings/GUIDialogAddonSettings.h"
-#include "ServiceBroker.h"
 #include "filesystem/Directory.h"
 #include "filesystem/ZipManager.h"
 #include "guilib/GUIWindowManager.h"
@@ -33,16 +32,17 @@
 #include "GUIUserMessages.h"
 #include "interfaces/generic/ScriptInvocationManager.h"
 #include "LangInfo.h"
-#include "log.h"
 #include "network/Network.h"
+#include "ServiceBroker.h"
 #include "settings/lib/Setting.h"
 #include "settings/Settings.h"
-#include "StringUtils.h"
-#include "URIUtils.h"
+#include "utils/log.h"
 #include "utils/POUtils.h"
+#include "utils/StringUtils.h"
 #include "utils/Temperature.h"
+#include "utils/URIUtils.h"
+#include "utils/XMLUtils.h"
 #include "XBDateTime.h"
-#include "XMLUtils.h"
 #ifdef TARGET_POSIX
 #include "platform/linux/XTimeUtils.h"
 #endif
